requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
filelock==3.7.0
fire==0.4.0
Flask==2.1.2
Flask-Cors==3.0.10
fugashi==1.1.2
huggingface-hub==0.7.0
idna==3.3
itsdangerous==2.1.2
jaconv==0.3
Jinja2==3.1.2
loguru==0.6.0
manga-ocr==0.1.7
MarkupSafe==2.1.1
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
pyperclip==1.8.2
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
six==1.16.0
termcolor==1.1.0
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.19.2
typing_extensions==4.2.0
unidic-lite==1.0.8
urllib3==1.26.9
Werkzeug==2.1.2
win32-setctime==1.1.0
